requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
black==22.8.0
blis==0.7.9
catalogue==2.0.8
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
confection==0.0.4
contourpy==1.0.6
croniter==1.3.8
cycler==0.11.0
cymem==2.0.7
et-xmlfile==1.1.0
exceptiongroup==1.0.4
Flask==2.2.2
fonttools==4.38.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.9
mypy-extensions==0.4.3
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
passlib==1.7.4
pathspec==0.10.1
pathy==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
preshed==3.0.8
prompt-toolkit==3.0.33
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.10.4
PyJWT==1.4.2
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
pytest==7.1.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
regex==2022.10.31
requests==2.28.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
srsly==2.4.5
thinc==8.1.7
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.7.0
types-pytz==2022.6.0.1
typing-extensions==4.3.0
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.8.1
